add server api
