+++ README.md
@@ -1,3 +1,7 @@
 # Dashboard-Pycharm
 code from pycharm
 This project just show kinds of chart with data.
+
+
+pip install -U django==1.9.10
+pip install pillow
